requirements.txt
streamlit
langchain
langchain-community
langchain-google-genai
langchain-chroma
langchain-huggingface
langgraph
pymupdf4llm
pymupdf
chromadb
tiktoken
python-dotenv
streamlit-pdf-viewer
sentence-transformers
pdfplumber